controllers complete

+++ Gallery.Api/Controllers/ExhibitController.cs
@@ -72,6 +72,9 @@ namespace Gallery.Api.Controllers
         public async Task<IActionResult> GetMine(CancellationToken ct)
         {
             var list = await _exhibitService.GetMineAsync(ct);
+            // add this user's permissions for each exhibit
+            AddPermissions(list);
+
             return Ok(list);
         }
 
@@ -89,7 +92,13 @@ namespace Gallery.Api.Controllers
         [SwaggerOperation(OperationId = "getUserExhibits")]
         public async Task<IActionResult> GetUserExhibits(Guid userId, CancellationToken ct)
         {
+            if (userId != User.GetId() && !await _authorizationService.AuthorizeAsync([SystemPermission.ViewExhibits], ct))
+                throw new ForbiddenException();
+
             var list = await _exhibitService.GetUserExhibitsAsync(userId, ct);
+            // add this user's permissions for each exhibit
+            AddPermissions(list);
+
             return Ok(list);
         }
 
@@ -107,6 +116,9 @@ namespace Gallery.Api.Controllers
         [SwaggerOperation(OperationId = "getCollectionExhibits")]
         public async Task<IActionResult> GetByCollection(Guid collectionId, CancellationToken ct)
         {
+            if (!await _authorizationService.AuthorizeAsync<Collection>(collectionId, [SystemPermission.ManageCollections], [CollectionPermission.ManageCollection], ct))
+                throw new ForbiddenException();
+
             var list = await _exhibitService.GetByCollectionAsync(collectionId, ct);
             return Ok(list);
         }
@@ -143,6 +155,9 @@ namespace Gallery.Api.Controllers
         [SwaggerOperation(OperationId = "getExhibit")]
         public async Task<IActionResult> Get(Guid id, CancellationToken ct)
         {
+            if (!await _authorizationService.AuthorizeAsync<Exhibit>(id, [SystemPermission.ViewExhibits], [ExhibitPermission.ViewExhibit], ct))
+                throw new ForbiddenException();
+
             var exhibit = await _exhibitService.GetAsync(id, ct);
 
             if (exhibit == null)
@@ -166,7 +181,9 @@ namespace Gallery.Api.Controllers
         [SwaggerOperation(OperationId = "createExhibit")]
         public async Task<IActionResult> Create([FromBody] Exhibit exhibit, CancellationToken ct)
         {
-            exhibit.CreatedBy = User.GetId();
+            if (!await _authorizationService.AuthorizeAsync([SystemPermission.CreateExhibits], ct))
+                throw new ForbiddenException();
+
             var createdExhibit = await _exhibitService.CreateAsync(exhibit, ct);
             return CreatedAtAction(nameof(this.Get), new { id = createdExhibit.Id }, createdExhibit);
         }
@@ -186,6 +203,10 @@ namespace Gallery.Api.Controllers
         [SwaggerOperation(OperationId = "copyExhibit")]
         public async Task<IActionResult> Copy(Guid id, CancellationToken ct)
         {
+            if (!await _authorizationService.AuthorizeAsync([SystemPermission.CreateExhibits], ct) ||
+               !await _authorizationService.AuthorizeAsync<Exhibit>(id, [SystemPermission.ViewExhibits], [ExhibitPermission.ViewExhibit], ct))
+                throw new ForbiddenException();
+
             var createdExhibit = await _exhibitService.CopyAsync(id, ct);
             return CreatedAtAction(nameof(this.Get), new { id = createdExhibit.Id }, createdExhibit);
         }
@@ -207,6 +228,9 @@ namespace Gallery.Api.Controllers
         [SwaggerOperation(OperationId = "updateExhibit")]
         public async Task<IActionResult> Update([FromRoute] Guid id, [FromBody] Exhibit exhibit, CancellationToken ct)
         {
+            if (!await _authorizationService.AuthorizeAsync<Exhibit>(id, [SystemPermission.EditExhibits], [ExhibitPermission.EditExhibit], ct))
+                throw new ForbiddenException();
+
             exhibit.ModifiedBy = User.GetId();
             var updatedExhibit = await _exhibitService.UpdateAsync(id, exhibit, ct);
             return Ok(updatedExhibit);
@@ -227,6 +251,9 @@ namespace Gallery.Api.Controllers
         [SwaggerOperation(OperationId = "setExhibitMoveAndInject")]
         public async Task<IActionResult> setExhibitMoveAndInject([FromRoute] Guid id, int move, int inject, CancellationToken ct)
         {
+            if (!await _authorizationService.AuthorizeAsync<Exhibit>(id, [SystemPermission.ManageExhibits], [ExhibitPermission.ManageExhibit], ct))
+                throw new ForbiddenException();
+
             var updatedExhibit = await _exhibitService.SetMoveAndInjectAsync(id, move, inject, ct);
             return Ok(updatedExhibit);
         }
@@ -246,6 +273,9 @@ namespace Gallery.Api.Controllers
         [SwaggerOperation(OperationId = "deleteExhibit")]
         public async Task<IActionResult> Delete(Guid id, CancellationToken ct)
         {
+            if (!await _authorizationService.AuthorizeAsync<Exhibit>(id, [SystemPermission.ManageExhibits], [ExhibitPermission.ManageExhibit], ct))
+                throw new ForbiddenException();
+
             await _exhibitService.DeleteAsync(id, ct);
             return NoContent();
         }
@@ -258,6 +288,9 @@ namespace Gallery.Api.Controllers
         [SwaggerOperation(OperationId = "uploadJsonFiles")]
         public async Task<IActionResult> UploadJsonAsync([FromForm] FileForm form, CancellationToken ct)
         {
+            if (!await _authorizationService.AuthorizeAsync([SystemPermission.CreateExhibits], ct))
+                throw new ForbiddenException();
+
             var result = await _exhibitService.UploadJsonAsync(form, ct);
             return Ok(result);
         }
@@ -270,6 +303,9 @@ namespace Gallery.Api.Controllers
         [SwaggerOperation(OperationId = "downloadJson")]
         public async Task<IActionResult> DownloadJsonAsync(Guid id, CancellationToken ct)
         {
+            if (!await _authorizationService.AuthorizeAsync<Exhibit>(id, [SystemPermission.ManageExhibits], [ExhibitPermission.ManageExhibit], ct))
+                throw new ForbiddenException();
+
             (var stream, var fileName) = await _exhibitService.DownloadJsonAsync(id, ct);
 
             // If this is wrapped in an Ok, it throws an exception
