+++ Gallery.Api/Controllers/TeamController.cs
@@ -6,8 +6,9 @@ using System.Collections.Generic;
 using System.Net;
 using System.Threading;
 using System.Threading.Tasks;
-using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Gallery.Api.Data.Enumerations;
+using Gallery.Api.Infrastructure.Authorization;
 using Gallery.Api.Infrastructure.Extensions;
 using Gallery.Api.Infrastructure.Exceptions;
 using Gallery.Api.Services;
@@ -19,30 +20,12 @@ namespace Gallery.Api.Controllers
     public class TeamController : BaseController
     {
         private readonly ITeamService _teamService;
-        private readonly IAuthorizationService _authorizationService;
+        private readonly IGalleryAuthorizationService _authorizationService;
 
-        public TeamController(ITeamService teamService, IAuthorizationService authorizationService)
+        public TeamController(ITeamService teamService, IGalleryAuthorizationService authorizationService)
         {
             _teamService = teamService;
             _authorizationService = authorizationService;
-        }
-
-        /// <summary>
-        /// Gets all Team in the system
-        /// </summary>
-        /// <remarks>
-        /// Returns a list of all of the Teams in the system.
-        /// <para />
-        /// Only accessible to a SuperUser
-        /// </remarks>
-        /// <returns></returns>
-        [HttpGet("teams")]
-        [ProducesResponseType(typeof(IEnumerable<Team>), (int)HttpStatusCode.OK)]
-        [SwaggerOperation(OperationId = "getTeams")]
-        public async Task<IActionResult> Get(CancellationToken ct)
-        {
-            var list = await _teamService.GetAsync(ct);
-            return Ok(list);
         }
 
         /// <summary>
@@ -57,41 +40,10 @@ namespace Gallery.Api.Controllers
         [SwaggerOperation(OperationId = "getMyExhibitTeams")]
         public async Task<IActionResult> GetMineByExhibit(Guid exhibitId, CancellationToken ct)
         {
+            if (!await _authorizationService.AuthorizeAsync<Exhibit>(exhibitId, [SystemPermission.ViewExhibits], [ExhibitPermission.ViewExhibit], ct))
+                throw new ForbiddenException();
+
             var list = await _teamService.GetMineByExhibitAsync(exhibitId, ct);
-            return Ok(list);
-        }
-
-        /// <summary>
-        /// Gets Teams for the specified user
-        /// </summary>
-        /// <remarks>
-        /// Returns a list of the specified user's Teams.
-        /// <para />
-        /// Only accessible to a SuperUser
-        /// </remarks>
-        /// <returns></returns>
-        [HttpGet("users/{userId}/teams")]
-        [ProducesResponseType(typeof(IEnumerable<Team>), (int)HttpStatusCode.OK)]
-        [SwaggerOperation(OperationId = "getTeamsByUser")]
-        public async Task<IActionResult> GetByUser([FromRoute] Guid userId, CancellationToken ct)
-        {
-            var list = await _teamService.GetByUserAsync(userId, ct);
-            return Ok(list);
-        }
-
-        /// <summary>
-        /// Gets Teams for the specified card
-        /// </summary>
-        /// <remarks>
-        /// Returns a list of the specified card's Teams.
-        /// </remarks>
-        /// <returns></returns>
-        [HttpGet("cards/{cardId}/teams")]
-        [ProducesResponseType(typeof(IEnumerable<Team>), (int)HttpStatusCode.OK)]
-        [SwaggerOperation(OperationId = "getTeamsByCard")]
-        public async Task<IActionResult> GetByCard([FromRoute] Guid cardId, CancellationToken ct)
-        {
-            var list = await _teamService.GetByCardAsync(cardId, ct);
             return Ok(list);
         }
 
@@ -107,6 +59,9 @@ namespace Gallery.Api.Controllers
         [SwaggerOperation(OperationId = "getTeamsByExhibit")]
         public async Task<IActionResult> GetByExhibit([FromRoute] Guid exhibitId, CancellationToken ct)
         {
+            if (!await _authorizationService.AuthorizeAsync<Exhibit>(exhibitId, [SystemPermission.EditExhibits], [ExhibitPermission.EditExhibit], ct))
+                throw new ForbiddenException();
+
             var list = await _teamService.GetByExhibitAsync(exhibitId, ct);
             return Ok(list);
         }
@@ -127,8 +82,10 @@ namespace Gallery.Api.Controllers
         [SwaggerOperation(OperationId = "getTeam")]
         public async Task<IActionResult> Get(Guid id, CancellationToken ct)
         {
-            var team = await _teamService.GetAsync(id, ct);
+            if (!await _authorizationService.AuthorizeAsync<Team>(id, [TeamPermission.ViewTeam], ct))
+                throw new ForbiddenException();
 
+            var team = await _teamService.GetAsync(id, ct);
             if (team == null)
                 throw new EntityNotFoundException<Team>();
 
@@ -150,7 +107,9 @@ namespace Gallery.Api.Controllers
         [SwaggerOperation(OperationId = "createTeam")]
         public async Task<IActionResult> Create([FromBody] Team team, CancellationToken ct)
         {
-            team.CreatedBy = User.GetId();
+            if (!await _authorizationService.AuthorizeAsync<Exhibit>(team.ExhibitId, [SystemPermission.EditExhibits], [ExhibitPermission.EditExhibit], ct))
+                throw new ForbiddenException();
+
             var createdTeam = await _teamService.CreateAsync(team, ct);
             return CreatedAtAction(nameof(this.Get), new { id = createdTeam.Id }, createdTeam);
         }
@@ -160,8 +119,6 @@ namespace Gallery.Api.Controllers
         /// </summary>
         /// <remarks>
         /// Updates a Team with the attributes specified
-        /// <para />
-        /// Accessible only to a SuperUser or a User on an Admin Team within the specified Team
         /// </remarks>
         /// <param name="id">The Id of the Exericse to update</param>
         /// <param name="team">The updated Team values</param>
@@ -171,7 +128,9 @@ namespace Gallery.Api.Controllers
         [SwaggerOperation(OperationId = "updateTeam")]
         public async Task<IActionResult> Update([FromRoute] Guid id, [FromBody] Team team, CancellationToken ct)
         {
-            team.ModifiedBy = User.GetId();
+            if (!await _authorizationService.AuthorizeAsync<Exhibit>(team.ExhibitId, [SystemPermission.EditExhibits], [ExhibitPermission.EditExhibit], ct))
+                throw new ForbiddenException();
+
             var updatedTeam = await _teamService.UpdateAsync(id, team, ct);
             return Ok(updatedTeam);
         }
@@ -191,6 +150,10 @@ namespace Gallery.Api.Controllers
         [SwaggerOperation(OperationId = "deleteTeam")]
         public async Task<IActionResult> Delete(Guid id, CancellationToken ct)
         {
+            var team = await _teamService.GetAsync(id, ct);
+            if (!await _authorizationService.AuthorizeAsync<Exhibit>(team.ExhibitId, [SystemPermission.EditExhibits], [ExhibitPermission.EditExhibit], ct))
+                throw new ForbiddenException();
+
             await _teamService.DeleteAsync(id, ct);
             return NoContent();
         }
