+++ Gallery.Api/Controllers/UserArticleController.cs
@@ -6,10 +6,11 @@ using System.Collections.Generic;
 using System.Net;
 using System.Threading;
 using System.Threading.Tasks;
-using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using Gallery.Api.Infrastructure.Extensions;
+using Gallery.Api.Data.Enumerations;
+using Gallery.Api.Infrastructure.Authorization;
 using Gallery.Api.Infrastructure.Exceptions;
+using Gallery.Api.Infrastructure.Extensions;
 using Gallery.Api.Services;
 using Gallery.Api.ViewModels;
 using Swashbuckle.AspNetCore.Annotations;
@@ -19,9 +20,9 @@ namespace Gallery.Api.Controllers
     public class UserArticleController : BaseController
     {
         private readonly IUserArticleService _userArticleService;
-        private readonly IAuthorizationService _authorizationService;
+        private readonly IGalleryAuthorizationService _authorizationService;
 
-        public UserArticleController(IUserArticleService userArticleService, IAuthorizationService authorizationService)
+        public UserArticleController(IUserArticleService userArticleService, IGalleryAuthorizationService authorizationService)
         {
             _userArticleService = userArticleService;
             _authorizationService = authorizationService;
@@ -31,7 +32,7 @@ namespace Gallery.Api.Controllers
         /// Gets UserArticles for an Exhibit
         /// </summary>
         /// <remarks>
-        /// Returns a list of UserArticles based on the exhibit's current move and current inject.
+        /// Returns a list of UserArticles for the exhibit.
         /// </remarks>
         /// <param name="exhibitId">The id of the Exhibit</param>
         /// <param name="ct"></param>
@@ -41,6 +42,9 @@ namespace Gallery.Api.Controllers
         [SwaggerOperation(OperationId = "getExhibitUserArticles")]
         public async Task<IActionResult> GetByExhibit(Guid exhibitId, CancellationToken ct)
         {
+            if (!await _authorizationService.AuthorizeAsync<Exhibit>(exhibitId, [SystemPermission.ViewExhibits], [ExhibitPermission.ViewExhibit], ct))
+                throw new ForbiddenException();
+
             var list = await _userArticleService.GetByExhibitAsync(exhibitId, ct);
             return Ok(list);
         }
@@ -60,6 +64,9 @@ namespace Gallery.Api.Controllers
         [SwaggerOperation(OperationId = "getExhibitTeamUserArticles")]
         public async Task<IActionResult> GetByExhibitTeam(Guid exhibitId, Guid teamId, CancellationToken ct)
         {
+            if (!await _authorizationService.AuthorizeAsync<Team>(teamId, [TeamPermission.ViewTeam], ct))
+                throw new ForbiddenException();
+
             var list = await _userArticleService.GetByExhibitTeamAsync(exhibitId, teamId, ct);
             return Ok(list);
         }
@@ -79,6 +86,9 @@ namespace Gallery.Api.Controllers
         [SwaggerOperation(OperationId = "getUnreadCount")]
         public async Task<IActionResult> GetUnreadCount(Guid exhibitId, Guid userId, CancellationToken ct)
         {
+            if (!await _authorizationService.AuthorizeAsync<Exhibit>(exhibitId, [SystemPermission.ViewExhibits], [ExhibitPermission.ViewExhibit], ct))
+                throw new ForbiddenException();
+
             var unreadArticles = await _userArticleService.GetUnreadCountAsync(exhibitId, userId, ct);
             return Ok(unreadArticles);
         }
@@ -88,8 +98,6 @@ namespace Gallery.Api.Controllers
         /// </summary>
         /// <remarks>
         /// Creates a new UserArticle with the attributes specified
-        /// <para />
-        /// Accessible only to a ContentDeveloper or an Administrator
         /// </remarks>
         /// <param name="userArticle">The data used to create the UserArticle</param>
         /// <param name="ct"></param>
@@ -98,7 +106,9 @@ namespace Gallery.Api.Controllers
         [SwaggerOperation(OperationId = "createUserArticle")]
         public async Task<IActionResult> Create([FromBody] UserArticle userArticle, CancellationToken ct)
         {
-            userArticle.CreatedBy = User.GetId();
+            if (!await _authorizationService.AuthorizeAsync<Exhibit>(userArticle.ExhibitId, [SystemPermission.EditExhibits], [ExhibitPermission.EditExhibit], ct))
+                throw new ForbiddenException();
+
             var createdUserArticle = await _userArticleService.CreateAsync(userArticle, ct);
             return Ok(createdUserArticle);
         }
@@ -117,6 +127,10 @@ namespace Gallery.Api.Controllers
         [SwaggerOperation(OperationId = "shareUserArticle")]
         public async Task<IActionResult> Share([FromRoute] Guid id, [FromBody] ShareDetails shareDetails, CancellationToken ct)
         {
+            var userArticle = await _userArticleService.GetAsync(id, ct);
+            if (!await _authorizationService.AuthorizeAsync<Exhibit>(userArticle.ExhibitId, [SystemPermission.EditExhibits], [ExhibitPermission.EditExhibit], ct))
+                throw new ForbiddenException();
+
             var sharedUserArticle = await _userArticleService.ShareAsync(id, shareDetails, ct);
             return Ok(sharedUserArticle);
         }
@@ -135,7 +149,11 @@ namespace Gallery.Api.Controllers
         [SwaggerOperation(OperationId = "setIsRead")]
         public async Task<IActionResult> SetIsRead([FromRoute] Guid id, [FromBody] bool isRead, CancellationToken ct)
         {
-            var userArticle = await _userArticleService.SetIsReadAsync(id, isRead, ct);
+            var userArticle = await _userArticleService.GetAsync(id, ct);
+            if (userArticle.UserId != User.GetId())
+                throw new ForbiddenException();
+
+            userArticle = await _userArticleService.SetIsReadAsync(id, isRead, ct);
             return Ok(userArticle);
         }
 
@@ -154,6 +172,10 @@ namespace Gallery.Api.Controllers
         [SwaggerOperation(OperationId = "deleteUserArticle")]
         public async Task<IActionResult> Delete(Guid id, CancellationToken ct)
         {
+            var userArticle = await _userArticleService.GetAsync(id, ct);
+            if (!await _authorizationService.AuthorizeAsync<Exhibit>(userArticle.ExhibitId, [SystemPermission.EditExhibits], [ExhibitPermission.EditExhibit], ct))
+                throw new ForbiddenException();
+
             await _userArticleService.DeleteAsync(id, ct);
             return NoContent();
         }
@@ -161,4 +183,3 @@ namespace Gallery.Api.Controllers
     }
 
 }
-
